backend/pyproject.toml
[project]
name = "vibecheck-backend"
version = "0.1.0"
description = "VibeCheck backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "anthropic>=0.25",
    "openai>=1.30",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "mcp>=1.0",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
